pyproject.toml
[tool.poetry]
name = "buttercup"
version = "0.24.0"
description = "The Discord Bot!"
authors = ["Grafeas Group Ltd. <devs@grafeas.org>"]
exclude = ['test', 'test.*', '*.test.*', '*.test']

[tool.poetry.dependencies]
"discord.py" = "^1.7.3"
discord-py-slash-command = "^3.0.3"
python = "^3.8.6"
toml = "^0.10.1"
blossom-wrapper = { git = "https://github.com/GrafeasGroup/blossom-wrapper.git", branch = "master" }
requests = "^2.25.1"
PyYAML = "^5.3.1"
python-dateutil = "^2.8.1"
asyncpraw = "^7.2.0"
matplotlib = "^3.4.2"
pandas = "^1.2.5"
seaborn = "^0.11.1"
shiv = "^1.0.1"

[tool.poetry.dev-dependencies]
better-exceptions = "^0.3.3"
black = "^22.6.0"
flake8 = { version = "^4.0.1", allow-prereleases = true }
flake8-annotations = "^2.0.1"
flake8-black = "^0.3.3"
flake8-docstrings = "^1.5.0"
flake8-variables-names = "^0.0.5"
pep8-naming = "^0.13.0"
pre-commit = "^2.4.0"
seed-isort-config = "^2.1.1"
isort = "^5.10.1"
pytest = "^7.1.2"
pytest-cov = "^3.0.0"
coveralls = "^3.2.0"
flake8-isort = "^4.1.2"

[tool.poetry.extras]
ci = ["pytest", "pytest-cov", "coveralls"]

[tool.isort]
profile = "black"
known_first_party = "src"
known_third_party = ["asyncpraw", "asyncprawcore", "blossom_wrapper", "dateutil", "discord", "discord_slash", "matplotlib", "pandas", "pytest", "pytz", "requests", "seaborn", "toml", "yaml"]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
